+++ back-end/lib/db.js
@@ -122,10 +122,10 @@ module.exports = {
           });
       });
     },
-    update: (id, user) => {
-      const original = store.users[id];
-      if (!original) throw Error("Unregistered user id");
-      store.users[id] = merge(original, user);
+    update: async (user) => {
+      if (!user.id) throw Error("Invalid user");
+      await db.put(`users:${user.id}`, JSON.stringify(user));
+      return merge(user, { id: user.id });
     },
     delete: (id, user) => {
       const original = store.users[id];